chore: refactor Misc test into AssertContains
