+++ tests/AssertContains.php
@@ -46,4 +46,17 @@ final class AssertContains extends TestCase
             ->contains('WIDGETBEF')
             ->contains('WIDGETAFT');
     }
+
+    // This is a legacy test for https://github.com/nunomaduro/laravel-mojito/issues/15
+    // because we don't use sprintf anymore it can be safely removed later on
+    public function testPercentageSymbolDoesNotBreakAssertions(): void
+    {
+        $html = file_get_contents(__DIR__.'/fixtures/button.html');
+        $this->expectException(AssertionFailedError::class);
+        $this->expectExceptionMessage(
+            "Failed asserting that the text `This text has a percentage 43% symbol` exists within `{$html}`"
+        );
+
+        $this->assertView('button')->contains('This text has a percentage 43% symbol');
+    }
 }
